# Replace `eval` by generated function