Benchmarked on Turing.jl/test/stdlib/RandomMeasures.jl

```
julia> @benchmark run_chain() # eval (before PR)
BenchmarkTools.Trial: 1 sample with 1 evaluation.
 Single result which took 15.231 s (1.52% GC) to evaluate,
 with a memory estimate of 679.26 MiB, over 9452913 allocations.

julia> @benchmark run_chain() # splatnew (after PR)
BenchmarkTools.Trial: 1 sample with 1 evaluation.
 Single result which took 14.665 s (1.55% GC) to evaluate,
 with a memory estimate of 655.23 MiB, over 8972863 allocations.

julia> @benchmark run_chain() # eval (before PR)
BenchmarkTools.Trial: 1 sample with 1 evaluation.
 Single result which took 16.465 s (1.41% GC) to evaluate,
 with a memory estimate of 679.21 MiB, over 9452903 allocations.

julia> @benchmark run_chain() # splatnew (after PR)
BenchmarkTools.Trial: 1 sample with 1 evaluation.
 Single result which took 15.409 s (1.50% GC) to evaluate,
 with a memory estimate of 655.34 MiB, over 8972879 allocations.
```

So about 3.5% runtime reduction and 5% allocations reduction on that test.

Would be good to run some more tests, but so far it looks promising.

ProfileSVG outputs look as follows.

**before PR**

![image](https://user-images.githubusercontent.com/20724914/154372964-3ea72be7-8300-4336-be3d-48d925f3f50b.png)

**After PR**

![image](https://user-images.githubusercontent.com/20724914/154373011-868bfdcc-e55d-4d05-9275-77f85638ac5d.png)