@rikhuijzer 2022-02-17
> And the number of calls to the __new__ function is 15k for those 500 iterations. Very reasonable trade off, right?

Based on the Action run times, I need to investigate further to double check this